% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Lasso_u}
\alias{B_T_Lasso_u}
\title{Bradley-Terry Weighted Likelihood Function with Lasso Peanlty's optimization}
\usage{
B_T_Lasso_u(df, ability, i0, lambda, wij, u)
}
\arguments{
\item{ability}{A column vector of teams ability, the last row is the home parameter whose rowname must be "at.home".}

\item{i0}{A teams index whose ability will be fixed as 0 (usually the team loss most).}

\item{lambda}{Lasso penalty of no statistical mean, a larger choice of lambda means higher penalty.
Usually the best penalty is chosen from Cross-Validation, where in-model's prediction power is maximized.}

\item{wij}{The weights added to the Lasso Penalty. Can be manually setted or determined using
function B_T_Wij}

\item{u}{The exponential decay rate}

\item{dataframe}{Dataframe with 5 columns. First column is the index of the home teams
(use numbers to denote teams).
Second column is the index of the away teams.
Third column is the number of wins of home teams (usually to be 0/1).
Fourth column is the number of wins of away teams (usually to be 0/1).
Fifth column is the scalar of time when the match is played until now (Time lag).}
}
\value{
Estimated abilities
}
\description{
Bradley-Terry Weighted Likelihood Function with Lasso Peanlty's optimization
}
