% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Step_1_2}
\alias{B_T_Step_1_2}
\title{Bradley-Terry Likelihood Function with Lasso Peanlty's optimization given fixed Lagraingian
and Quadratic Penalty}
\usage{
B_T_Step_1_2(df, ability, i0, wij, uij, theta, n, v, lambda)
}
\description{
Bradley-Terry Likelihood Function with Lasso Peanlty's optimization given fixed Lagraingian
and Quadratic Penalty
}
