% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Step_1_2_u}
\alias{B_T_Step_1_2_u}
\title{Bradley-Terry Weighted Likelihood Function with Lasso Peanlty's optimization given fixed Lagraingian
and Quadratic Penalty}
\usage{
B_T_Step_1_2_u(df, ability, i0, wij, uij, theta, n, v, lambda, u)
}
\description{
Bradley-Terry Weighted Likelihood Function with Lasso Peanlty's optimization given fixed Lagraingian
and Quadratic Penalty
}
