% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Step_1_2_u_fhome}
\alias{B_T_Step_1_2_u_fhome}
\title{Bradley-Terry Weighted Likelihood Function with fixed home parameter and Lasso Peanlty's optimization given fixed Lagraingian
and Quadratic Penalty}
\usage{
B_T_Step_1_2_u_fhome(df, ability, i0, wij, uij, theta, n, v, lambda, u, home)
}
\description{
Bradley-Terry Weighted Likelihood Function with fixed home parameter and Lasso Peanlty's optimization given fixed Lagraingian
and Quadratic Penalty
}
