% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Weighted_fp}
\alias{B_T_Weighted_fp}
\title{Bradley-Terry Weighted Likelihood Function with home parameter for Likelihood Ratio Test of
abilities time constancy}
\usage{
B_T_Weighted_fp(df, ability, uf, u, i0, date)
}
\arguments{
\item{df}{Dataframe with 5 columns. First column is the index of the home teams
(use numbers to denote teams).
Second column is the index of the away teams.
Third column is the number of wins of home teams (usually to be 0/1).
Fourth column is the number of wins of away teams (usually to be 0/1).
Fifth column is the scalar of time when the match is played until now (Time lag).}

\item{ability}{A column vector of teams ability, the last row is the home parameter  whose rowname must be "at.home".
The row number is consistent with the team's index shown in dataframe. Column name must be "xn".}

\item{uf}{Team index whose time constancy will be tested.}

\item{u}{The exponential decay rate.}

\item{i0}{is a teams index whose ability will be fixed as 0 (usually the team loss most).}

\item{date}{the date were selected team will have different abilities before and after
(usually chosen to be the mid time when abilities are expected to change).}
}
\value{
The estimated abilities.
}
\description{
Bradley-Terry Weighted Likelihood Function with home parameter for Likelihood Ratio Test of
abilities time constancy
}
