% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Weighted_Lasso_R}
\alias{B_T_Weighted_Lasso_R}
\title{Bradley-Terry Weighted Likelihood function with and Lasso Peanlty's optimization where Wij are setting to be 1
Prediction on one future period is done}
\usage{
B_T_Weighted_Lasso_R(df, ability, i0, date, u, lambda, dd)
}
\arguments{
\item{ability}{A column vector of teams ability, the last row is the home parameter whose rowname must be "at.home".}

\item{i0}{A teams index whose ability will be fixed as 0 (usually the team loss most).}

\item{date}{A vector which separate all rows of datasets into many parts.}

\item{u}{The exponential decay rate.}

\item{lambda}{Lasso penalty of no statistical mean, a larger choice of lambda means higher penalty.
Usually the best penalty is chosen from Cross-Validation, where in-model's prediction power is maximized.}

\item{dd}{The index of one value from date. Trained model will be run before this date and prediction will be done
on the previous part of dataset.}

\item{dataframe}{Dataframe with 5 columns. First column is the index of the home teams
(use numbers to denote teams).
Second column is the index of the away teams.
Third column is the number of wins of home teams (usually to be 0/1).
Fourth column is the number of wins of away teams (usually to be 0/1).
Fifth column is the scalar of time when the match is played until now (Time lag).}
}
\value{
The estimated abilities and predicted likelihood
}
\description{
Bradley-Terry Weighted Likelihood function with and Lasso Peanlty's optimization where Wij are setting to be 1
Prediction on one future period is done
}
