% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Weighted_Lasso_R_wij_fhome}
\alias{B_T_Weighted_Lasso_R_wij_fhome}
\title{Bradley-Terry Weighted Likelihood Estimation with fixed home parameter and Lasso Peanlty's optimization}
\usage{
B_T_Weighted_Lasso_R_wij_fhome(df, ability, i0, u, penalty, uk, home)
}
\arguments{
\item{ability}{A column vector of teams ability, the last row is the home parameter whose rowname must be "at.home".}

\item{i0}{A teams index whose ability will be fixed as 0 (usually the team loss most).}

\item{u}{The exponential decay rate.}

\item{penalty}{Penalty level = s/max(s)}

\item{uk}{Exponetial rate between lambda and true penalty level (usually setting to be -0.642)}

\item{home}{Fixed home ability}

\item{dataframe}{Dataframe with 5 columns. First column is the index of the home teams
(use numbers to denote teams).
Second column is the index of the away teams.
Third column is the number of wins of home teams (usually to be 0/1).
Fourth column is the number of wins of away teams (usually to be 0/1).
Fifth column is the scalar of time when the match is played until now (Time lag).}

\item{lambda}{Lasso penalty of no statistical mean, a larger choice of lambda means higher penalty.
Usually the best penalty is chosen from Cross-Validation, where in-model's prediction power is maximized.}
}
\value{
The estimated abilities
}
\description{
Bradley-Terry Weighted Likelihood Estimation with fixed home parameter and Lasso Peanlty's optimization
}
