% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Weighted_Lasso_TR}
\alias{B_T_Weighted_Lasso_TR}
\title{Bradley-Terry Weighted Likelihood Estimation before a specified time
Prediction on one future period is done}
\usage{
B_T_Weighted_Lasso_TR(df, ability, i0, date, u, dd)
}
\description{
Bradley-Terry Weighted Likelihood Estimation before a specified time
Prediction on one future period is done
}
