% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_Wij_u_fhome}
\alias{B_T_Wij_u_fhome}
\title{Weighted Adapted Lasso's weight determination function with fixed home paramter}
\usage{
B_T_Wij_u_fhome(df, ability, i0, n, u, home)
}
\description{
Weighted Adapted Lasso's weight determination function with fixed home paramter
}
