% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_with_Qua}
\alias{B_T_with_Qua}
\title{Bradely-Terry Model Likelihood function with Lagrangian and Quadratic Penalty}
\usage{
B_T_with_Qua(df, ability, i0, theta, n, v, uij)
}
\description{
Bradely-Terry Model Likelihood function with Lagrangian and Quadratic Penalty
}
