% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_with_Qua_u}
\alias{B_T_with_Qua_u}
\title{Bradely-Terry Model Weighted Likelihood function with Lagrangian and Quadratic Penalty}
\usage{
B_T_with_Qua_u(df, ability, i0, theta, n, v, uij, u)
}
\description{
Bradely-Terry Model Weighted Likelihood function with Lagrangian and Quadratic Penalty
}
