% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{B_T_with_Qua_u_fhome}
\alias{B_T_with_Qua_u_fhome}
\title{Bradely-Terry Model Weighted Likelihood function with fixed home parameter, Lagrangian and Quadratic Penalty}
\usage{
B_T_with_Qua_u_fhome(df, ability, i0, theta, n, v, uij, u, home)
}
\description{
Bradely-Terry Model Weighted Likelihood function with fixed home parameter, Lagrangian and Quadratic Penalty
}
