% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Likelihood}
\alias{Likelihood}
\title{Likelihood computing Function on Bradley-Terry Likelihood and Lasso penalty Part}
\usage{
Likelihood(df, ability, wij, lambda, theta, n)
}
\description{
Likelihood computing Function on Bradley-Terry Likelihood and Lasso penalty Part
}
