% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Likelihood_All}
\alias{Likelihood_All}
\title{Likelihood computing Function on Bradley-Terry Likelihood and all penalty Part}
\usage{
Likelihood_All(df, ability, theta, n, v, wij, uij, lambda)
}
\description{
Likelihood computing Function on Bradley-Terry Likelihood and all penalty Part
}
