% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Likelihood_All_u}
\alias{Likelihood_All_u}
\title{Likelihood computing Function on Weighted Bradley-Terry Likelihood and all penalty Part}
\usage{
Likelihood_All_u(df, ability, theta, n, v, wij, uij, lambda, u)
}
\description{
Likelihood computing Function on Weighted Bradley-Terry Likelihood and all penalty Part
}
