% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Likelihood_Pure}
\alias{Likelihood_Pure}
\title{Likelihood computing Function on Bradley-Terry Likelihood only}
\usage{
Likelihood_Pure(df, ability, n)
}
\description{
Likelihood computing Function on Bradley-Terry Likelihood only
}
