% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Likelihood_Pure_u}
\alias{Likelihood_Pure_u}
\title{Likelihood computing Function on weighted Bradley-Terry Likelihood only}
\usage{
Likelihood_Pure_u(df, ability, n, u)
}
\description{
Likelihood computing Function on weighted Bradley-Terry Likelihood only
}
