% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Likelihood_Qua}
\alias{Likelihood_Qua}
\title{Likelihood computing Function on Bradley-Terry Likelihood and Quadratic penalty Part}
\usage{
Likelihood_Qua(df, ability, theta, n, v)
}
\description{
Likelihood computing Function on Bradley-Terry Likelihood and Quadratic penalty Part
}
