% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Likelihood_Qua_lambda}
\alias{Likelihood_Qua_lambda}
\title{Likelihood computing Function on Bradley-Terry Likelihood and Lasso and Quadratic penalty Part}
\usage{
Likelihood_Qua_lambda(df, ability, theta, n, v, wij, lambda)
}
\description{
Likelihood computing Function on Bradley-Terry Likelihood and Lasso and Quadratic penalty Part
}
