% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{MSE_v2}
\alias{MSE_v2}
\title{MSE of linearly changed ability with two teams involved and no home effect}
\usage{
MSE_v2(a, n, a1, k1, p)
}
\arguments{
\item{a}{Exponetial decay rate.}

\item{n}{Number of matched.}

\item{a1}{The ability score at the beginning.}

\item{k1}{Slope of linearly changed ability}

\item{p}{Lasso penalty level = s/max(s)}
}
\value{
Bias, Variance and MSE
}
\description{
MSE of linearly changed ability with two teams involved and no home effect
}
