% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Time_table2}
\alias{Time_table2}
\title{Time Table construction for two teams with home effect}
\usage{
Time_table2(n, x1, x2, home)
}
\description{
Time Table construction for two teams with home effect
}
