% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection of Functions.R
\name{Time_table4}
\alias{Time_table4}
\title{Time Table construction for four teams with home effect}
\usage{
Time_table4(n, x1, x2, x3, x4, home)
}
\description{
Time Table construction for four teams with home effect
}
